refactor: #18 Content 삭제 및 메인, 메뉴 컴포넌트 props 제거
 - 이유: 어차피 Content는 Recoil 사용해서 굳이 props 로 넘겨줄 필요 X.
