+++ src/components/meeteam/menu/Menu.tsx
@@ -1,23 +1,28 @@
 import React from 'react';
 import S from './Menu.styled';
+import { useRecoilState } from 'recoil';
+import { contentState } from '../../../atom';
 
-const Menu = (props: {
-	menu: string;
-	clickedHandler: React.MouseEventHandler<HTMLButtonElement>;
-}) => {
+const Menu = () => {
+	const [content, setContent] = useRecoilState(contentState);
+
+	const clickedHandler: React.MouseEventHandler<HTMLButtonElement> = e => {
+		const button = e.target as HTMLButtonElement;
+		if (!button.textContent) {
+			throw new Error('No Content');
+		}
+		setContent(button.textContent);
+	};
+
 	const menuList = ['대시보드', '밋팀', '멤버', '구인 현황', '설정'] as const;
 	type menuType = (typeof menuList)[number];
 
-	const checkColorChange = (currentMenu: menuType) => currentMenu === props.menu;
+	const checkColorChange = (currentMenu: menuType) => currentMenu === content;
 
 	return (
 		<S.MenuRow>
 			{menuList.map((currentMenu, index) => (
-				<S.MenuBox
-					onClick={props.clickedHandler}
-					$color={checkColorChange(currentMenu)}
-					key={index}
-				>
+				<S.MenuBox onClick={clickedHandler} $color={checkColorChange(currentMenu)} key={index}>
 					{currentMenu}
 				</S.MenuBox>
 			))}
